Fixed some 64bit compatibility issues
git-svn-id: http://svn.code.sf.net/p/angelscript/code/trunk@362 404ce1b2-830e-0410-a2e2-b09542c77caf

+++ sdk/tests/test_feature/source/test_return_with_cdecl_objfirst.cpp
@@ -11,22 +11,22 @@
 class Class1
 {
 public:
-	unsigned long a;
+	unsigned int a;
 };
 
 class Class2
 {
 public:
-	unsigned long a;
-	unsigned long b;
+	unsigned int a;
+	unsigned int b;
 };
 
 class Class3
 {
 public:
-	unsigned long a;
-	unsigned long b;
-	unsigned long c;
+	unsigned int a;
+	unsigned int b;
+	unsigned int c;
 };
 
 class ClassNotComplex {};
